backend/python_django/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "tweety"
version = "0.1.0"
description = "A Twitter clone web app"
authors = ["Ahmad <zoaby.am@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.6"
django-environ = "0.4.5"
django-extensions = "3.1.3"
Pillow = "^8.3.2"
psycopg2-binary = "^2.9.1"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"
django-oauth-toolkit = "^1.5.0"
drf-nested-routers = "^0.93.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-django = "^4.4.0"
pytest-sugar = "^0.9.4"
pytest-cov = "^2.12.1"
coverage = "^5.5"
flake8 = "^3.9.2"
flake8-bugbear = "^21.4.3"
flake8-bandit = "^2.1.2"
pep8-naming = "^0.12.1"
mypy = "^0.910"
black = "^21.7b0"
isort = "^5.9.3"
factory-boy = "^3.2.0"
django-debug-toolbar = "^3.2.2"
django-stubs = "^1.8.0"
dennis = "^0.9"
dotenv-linter = "^0.2.0"
nplusone = "^1.0.0"
safety = "^1.10.3"
Werkzeug = "^2.0.2"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-s -vv --cov --cov-report=html --ds=config.settings --reuse-db --cache-clear --disable-pytest-warnings"
python_files = "tests test_*.py"
norecursedirs = ".tox .git */migrations/* */static/* docs venv"


[tool.coverage.run]
omit = [".*", "*/site-packages/*", "*/tests/*", "*/config/*"]

[tool.coverage.report]
fail_under = 90


[tool.isort]
include_trailing_comma = true
line_length = 100
multi_line_output = 3
use_parentheses = true
skip = "migrations"

[tool.black]
line-length = 100
exclude = '''
/(
  | migrations
)/

'''
